release test bat file check

+++ buildhelper.ps1
@@ -11,6 +11,7 @@ $temp = [System.IO.Path]::GetTempPath();
 $outfile = "$($temp)buildoutput.log";
 $fipoutfile = "$($temp)fipbuildoutput.log";
 $cxoutfile = "$($temp)cxbuildoutput.log";
+$migrationlog = "c:\ProgramData\Fieldpro\FIP_SYSTEM_LOG.LOG";
 $svc = New-WebServiceProxy –Uri ‘http://192.168.0.1/taskmanagerbeta/trservice.asmx?WSDL’
 #$svc = New-WebServiceProxy –Uri ‘http://localhost:8311/TRService.asmx?WSDL’
 $request = $null;
@@ -168,6 +169,14 @@ function Invoke-CodeBuilder()
     $testcmd = "RELEASE_TEST.BAT $($user) $($version) $($ttid) $($comment) $($vspath)";
     Write-State "$($testcmd)"
     cmd /c "$($testcmd)" | Out-File $($outfile) -Append;
+    if ($LASTEXITCODE -eq 1) {
+        Write-State "Failed to run release test."
+        Copy-Item $migrationlog -Destination "$($pathtolog)$($request.ID).log"
+        $svc.FailBuild($request.ID)
+        stop-computer
+        exit
+    }
+    
     Stop-Service "MSSQLSERVER"
 
     $fileerror = Select-String -Path $outfile -Pattern "^Error:" #line starts with 'error:'
